% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_process_response.R
\name{fdic_process_response}
\alias{fdic_process_response}
\title{Convert FDIC Response to Dataframe}
\usage{
fdic_process_response(fdic_resp)
}
\arguments{
\item{fdic_resp}{response from call_fdic_api()}
}
\value{
A dataframe
}
\description{
Function to convert response from call_fdic_api() to a dataframe
}
\examples{
resp <- call_fdic_api(api = "financials", filters = "RSSDID: 37", fields = "RSSDID,REPDTE,ASSET,DEP", limit = 10)
}
